RF: Move commands to submodule
