+++ fendermint/app/src/cmd/mod.rs
@@ -1,0 +1,6 @@
+// Copyright 2022-2023 Protocol Labs
+// SPDX-License-Identifier: Apache-2.0, MIT
+
+//! CLI command implementations.
+
+pub mod run;
