+++ fendermint/app/src/lib.rs
@@ -1,8 +1,6 @@
 // Copyright 2022-2023 Protocol Labs
 // SPDX-License-Identifier: Apache-2.0, MIT
 mod app;
-pub mod options;
-pub mod settings;
 mod store;
 mod tmconv;
