+++ fendermint/app/src/main.rs
@@ -2,18 +2,13 @@
 // SPDX-License-Identifier: Apache-2.0, MIT
 
 use clap::Parser;
-use fendermint_abci::ApplicationService;
-use fendermint_app::{
-    options::{Command, Options},
-    settings::Settings,
-    App, AppStore,
-};
-use fendermint_rocksdb::{RocksDb, RocksDbConfig};
-use fendermint_vm_interpreter::{
-    bytes::BytesMessageInterpreter, chain::ChainMessageInterpreter, fvm::FvmMessageInterpreter,
-    signed::SignedMessageInterpreter,
-};
 use tracing_subscriber::FmtSubscriber;
+
+mod cmd;
+mod options;
+mod settings;
+
+use options::{Commands, Options};
 
 #[tokio::main]
 async fn main() {
@@ -26,60 +21,15 @@ async fn main() {
 
     tracing::subscriber::set_global_default(subscriber).expect("setting default subscriber failed");
 
-    if let Some(Command::Run { ref mode }) = opts.command {
-        let config_dir = match opts.config_dir() {
-            Some(d) if d.is_dir() => d,
-            Some(d) if d.exists() => panic!("config '{d:?}' is a not a directory"),
-            Some(d) => panic!("config '{d:?}' does not exist"),
-            None => panic!("could not find a config directory to use"),
+    if let Some(ref cmd) = opts.command {
+        let result = match cmd {
+            Commands::Run { mode } => cmd::run::run(opts.config_dir(), mode.as_ref()).await,
         };
-
-        let settings = Settings::new(config_dir, mode).expect("error parsing settings");
-
-        let interpreter = FvmMessageInterpreter::<RocksDb>::new();
-        let interpreter = SignedMessageInterpreter::new(interpreter);
-        let interpreter = ChainMessageInterpreter::new(interpreter);
-        let interpreter = BytesMessageInterpreter::new(interpreter);
-
-        let db = open_db(&settings).expect("error opening DB");
-        let app_ns = db.new_cf_handle("app").unwrap();
-        let state_hist_ns = db.new_cf_handle("state_hist").unwrap();
-
-        let app = App::<_, AppStore, _>::new(
-            db,
-            settings.builtin_actors_bundle,
-            app_ns,
-            state_hist_ns,
-            interpreter,
-        );
-
-        let service = ApplicationService(app);
-
-        // Split it into components.
-        let (consensus, mempool, snapshot, info) =
-            tower_abci::split::service(service, settings.abci.bound);
-
-        // Hand those components to the ABCI server. This is where tower layers could be added.
-        let server = tower_abci::Server::builder()
-            .consensus(consensus)
-            .snapshot(snapshot)
-            .mempool(mempool)
-            .info(info)
-            .finish()
-            .expect("error creating ABCI server");
-
-        // Run the ABCI server.
-        server
-            .listen(settings.abci.listen_addr())
-            .await
-            .expect("error listening to ABCI requests");
+        if let Err(e) = result {
+            tracing::error!("failed to execute {cmd:?}: {e}");
+            std::process::exit(1);
+        }
     }
-}
-
-fn open_db(settings: &Settings) -> anyhow::Result<RocksDb> {
-    let path = settings.data_dir.join("rocksdb");
-    let db = RocksDb::open(path, &RocksDbConfig::default())?;
-    Ok(db)
 }
 
 #[cfg(test)]
